fix(go/plugins/vertexai/modelgarden): correct Llama 3.3 caps, share env helper, guard Anthropic.DefineModel
- Llama 3.3 70B Instruct is text-only on Vertex MaaS; switch its
  Supports to BasicText. Llama 4 Maverick/Scout remain Multimodal.
- Centralise resolveVertexMaasEnv in models.go and use it from
  anthropic.go, llama.go, and mistral.go (Anthropic's inline ~20-line
  block collapses to one call).
- Anthropic.DefineModel now acquires the mutex and checks initted,
  matching Llama/Mistral. Calling DefineModel before Init returned a
  zero-valued client; it now returns a "not initialized" error.
- Add define_model_test.go covering DefineModel-before-Init for all
  three plugins.
